Fix BL32 to boot xipImage from MRAM too.
This is a fix for ARM TF (BL32), when the Linux kernel xipImage is to be
placed on MRAM.

Initially ARM TF expected the Linux kernel xipImage to be on the NOR
flash at 0xD000_0000 in case of Dev board or at 0xE000_0000 in case of
CARRIER board.

The current fix allows the Linux kernel to be placed on the MRAM and
also fixes the SE_UART PinMUX issue.

Reviewed-by: Jagadeesh Krishnanjanappa <jagadeesh@alifsemi.com>
Signed-off-by: Manjunath S Matti <manjunath.matti@alifsemi.com>